Started work on implementing the Database module

+++ server/src/com/gdx/mirage/server/record/OldMapRec.java
@@ -6,17 +6,17 @@ public class OldMapRec {
 
     public String name;
     public long revision;
-    public byte moral;
+    public short moral;
     public int up;
     public int down;
     public int left;
     public int right;
-    public byte music;
+    public short music;
     public int bootMap;
-    public byte bootX;
-    public byte bootY;
-    public byte shop;
+    public short bootX;
+    public short bootY;
+    public short shop;
     public final TileRec[][] tile = new TileRec[Constants.MAX_MAPX][Constants.MAX_MAPY];
-    public final byte[] npc = new byte[Constants.MAX_MAP_NPCS];
+    public final short[] npc = new short[Constants.MAX_MAP_NPCS];
 
 }
